requirements.txt
Pillow==10.3.0
ImageHash==4.3.1
Send2Trash==1.8.3
